# docs/api-gateway/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway for FCH messenger. All protected routes require JWT cookie.",
        "title": "FCH API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/chat/{chatID}/send": {
            "get": {
                "description": "Renders chat page for a specific chat\nRenders list of user's chats\nCreates or retrieves a personal chat with another user\nGroup chat creation page (GET) or submission (POST)\nJoin a group chat via invite code\nUpgrades to WebSocket for real-time messaging",
                "produces": [
                    "text/html",
                    "text/html",
                    "text/html",
                    "text/html",
                    "text/html"
                ],
                "tags": [
                    "chat",
                    "chat",
                    "chat",
                    "chat",
                    "chat",
                    "chat"
                ],
                "summary": "WebSocket chat connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group name (POST only)",
                        "name": "group_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Invite code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Chat ID",
                        "name": "chatID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching protocols to WebSocket"
                    },
                    "200": {
                        "description": "Group create form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "302": {
                        "description": "Redirect to /chat/{groupID}"
                    },
                    "400": {
                        "description": "Invalid or expired invite code"
                    },
                    "401": {
                        "description": "Unauthorized — missing or invalid JWT"
                    },
                    "403": {
                        "description": "Access denied — user not in chat"
                    }
                }
            }
        },
        "/chat/{userID}": {
            "get": {
                "description": "Renders chat page for a specific chat\nRenders list of user's chats\nCreates or retrieves a personal chat with another user\nGroup chat creation page (GET) or submission (POST)\nJoin a group chat via invite code\nUpgrades to WebSocket for real-time messaging",
                "produces": [
                    "text/html",
                    "text/html",
                    "text/html",
                    "text/html",
                    "text/html"
                ],
                "tags": [
                    "chat",
                    "chat",
                    "chat",
                    "chat",
                    "chat",
                    "chat"
                ],
                "summary": "WebSocket chat connection",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Chat ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Target user ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group name (POST only)",
                        "name": "group_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Invite code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching protocols to WebSocket"
                    },
                    "200": {
                        "description": "Group create form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "302": {
                        "description": "Redirect to /chat/{groupID}"
                    },
                    "400": {
                        "description": "Invalid or expired invite code"
                    },
                    "401": {
                        "description": "Unauthorized — missing or invalid JWT"
                    },
                    "403": {
                        "description": "Access denied — user not in chat"
                    }
                }
            }
        },
        "/group-chat/create": {
            "get": {
                "description": "Renders chat page for a specific chat\nRenders list of user's chats\nCreates or retrieves a personal chat with another user\nGroup chat creation page (GET) or submission (POST)\nJoin a group chat via invite code\nUpgrades to WebSocket for real-time messaging",
                "produces": [
                    "text/html",
                    "text/html",
                    "text/html",
                    "text/html",
                    "text/html"
                ],
                "tags": [
                    "chat",
                    "chat",
                    "chat",
                    "chat",
                    "chat",
                    "chat"
                ],
                "summary": "WebSocket chat connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group name (POST only)",
                        "name": "group_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Invite code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching protocols to WebSocket"
                    },
                    "200": {
                        "description": "Group create form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "302": {
                        "description": "Redirect to /chat/{groupID}"
                    },
                    "400": {
                        "description": "Invalid or expired invite code"
                    },
                    "401": {
                        "description": "Unauthorized — missing or invalid JWT"
                    },
                    "403": {
                        "description": "Access denied — user not in chat"
                    }
                }
            },
            "post": {
                "description": "Renders chat page for a specific chat\nRenders list of user's chats\nCreates or retrieves a personal chat with another user\nGroup chat creation page (GET) or submission (POST)\nJoin a group chat via invite code\nUpgrades to WebSocket for real-time messaging",
                "produces": [
                    "text/html",
                    "text/html",
                    "text/html",
                    "text/html",
                    "text/html"
                ],
                "tags": [
                    "chat",
                    "chat",
                    "chat",
                    "chat",
                    "chat",
                    "chat"
                ],
                "summary": "WebSocket chat connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group name (POST only)",
                        "name": "group_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Invite code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching protocols to WebSocket"
                    },
                    "200": {
                        "description": "Group create form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "302": {
                        "description": "Redirect to /chat/{groupID}"
                    },
                    "400": {
                        "description": "Invalid or expired invite code"
                    },
                    "401": {
                        "description": "Unauthorized — missing or invalid JWT"
                    },
                    "403": {
                        "description": "Access denied — user not in chat"
                    }
                }
            }
        },
        "/group-chat/join-to-chat/{groupID}": {
            "get": {
                "description": "Renders chat page for a specific chat\nRenders list of user's chats\nCreates or retrieves a personal chat with another user\nGroup chat creation page (GET) or submission (POST)\nJoin a group chat via invite code\nUpgrades to WebSocket for real-time messaging",
                "produces": [
                    "text/html",
                    "text/html",
                    "text/html",
                    "text/html",
                    "text/html"
                ],
                "tags": [
                    "chat",
                    "chat",
                    "chat",
                    "chat",
                    "chat",
                    "chat"
                ],
                "summary": "WebSocket chat connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group name (POST only)",
                        "name": "group_name",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Group chat ID",
                        "name": "groupID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Invite code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching protocols to WebSocket"
                    },
                    "200": {
                        "description": "Group create form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "302": {
                        "description": "Redirect to /chat/{groupID}"
                    },
                    "400": {
                        "description": "Invalid or expired invite code"
                    },
                    "401": {
                        "description": "Unauthorized — missing or invalid JWT"
                    },
                    "403": {
                        "description": "Access denied — user not in chat"
                    }
                }
            }
        },
        "/login": {
            "get": {
                "description": "Login page (GET) or authentication (POST)\nRegistration page (GET) or submission (POST)\nSearch users by username",
                "produces": [
                    "text/html",
                    "text/html",
                    "text/html"
                ],
                "tags": [
                    "user",
                    "user",
                    "user"
                ],
                "summary": "Search users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username (POST only)",
                        "name": "username",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Password (POST only)",
                        "name": "password",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Username (POST only)",
                        "name": "username",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Password (POST only)",
                        "name": "password",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results page",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body"
                    },
                    "401": {
                        "description": "Unauthorized — missing or invalid JWT"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "409": {
                        "description": "User already exists"
                    }
                }
            },
            "post": {
                "description": "Login page (GET) or authentication (POST)\nRegistration page (GET) or submission (POST)\nSearch users by username",
                "produces": [
                    "text/html",
                    "text/html",
                    "text/html"
                ],
                "tags": [
                    "user",
                    "user",
                    "user"
                ],
                "summary": "Search users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username (POST only)",
                        "name": "username",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Password (POST only)",
                        "name": "password",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Username (POST only)",
                        "name": "username",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Password (POST only)",
                        "name": "password",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results page",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body"
                    },
                    "401": {
                        "description": "Unauthorized — missing or invalid JWT"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "409": {
                        "description": "User already exists"
                    }
                }
            }
        },
        "/my-chats": {
            "get": {
                "description": "Renders chat page for a specific chat\nRenders list of user's chats\nCreates or retrieves a personal chat with another user\nGroup chat creation page (GET) or submission (POST)\nJoin a group chat via invite code\nUpgrades to WebSocket for real-time messaging",
                "produces": [
                    "text/html",
                    "text/html",
                    "text/html",
                    "text/html",
                    "text/html"
                ],
                "tags": [
                    "chat",
                    "chat",
                    "chat",
                    "chat",
                    "chat",
                    "chat"
                ],
                "summary": "WebSocket chat connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group name (POST only)",
                        "name": "group_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Invite code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching protocols to WebSocket"
                    },
                    "200": {
                        "description": "Group create form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "302": {
                        "description": "Redirect to /chat/{groupID}"
                    },
                    "400": {
                        "description": "Invalid or expired invite code"
                    },
                    "401": {
                        "description": "Unauthorized — missing or invalid JWT"
                    },
                    "403": {
                        "description": "Access denied — user not in chat"
                    }
                }
            }
        },
...
        "/search/{username}": {
            "get": {
                "description": "Login page (GET) or authentication (POST)\nRegistration page (GET) or submission (POST)\nSearch users by username",
                "produces": [
                    "text/html",
                    "text/html",
                    "text/html"
                ],
                "tags": [
                    "user",
                    "user",
                    "user"
                ],
                "summary": "Search users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username (POST only)",
                        "name": "username",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Password (POST only)",
                        "name": "password",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Username (POST only)",
                        "name": "username",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Password (POST only)",
                        "name": "password",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Username to search",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results page",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body"
                    },
                    "401": {
                        "description": "Unauthorized — missing or invalid JWT"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "409": {
                        "description": "User already exists"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "jwt",
            "in": "cookie"
        }
    }
}